+++ bricks/clean_feature/__brick__/lib/domain/repositories/{{name.snakeCase()}}_repository.dart
@@ -1,6 +1,0 @@
-import '../../app/utilities/typedef.dart';
-import '../entities/{{name.snakeCase()}}_entity.dart';
-
-abstract class {{name.pascalCase()}}Repository {
-  ResultFuture<{{name.pascalCase()}}Entity> get{{name.pascalCase()}}();
-}