+++ bricks/project_setup/__brick__/lib/app/constants/api_constants.dart
@@ -1,8 +1,0 @@
-import 'package:flutter_dotenv/flutter_dotenv.dart';
-
-abstract class ApiConstants {
-  static String kBaseUrl = dotenv.env['BASE_URL'] ?? '';
-
-  // user
-  static String kUserProfile = 'user_profile';
-}