+++ bricks/project_setup/__brick__/lib/app/routing/route_constants.dart
@@ -1,4 +1,0 @@
-class RouteConstants {
-  static const String kSplash = '/splash';
-  static const String kOnBoarding = '/on-boarding';
-}