+++ bricks/project_setup/__brick__/lib/src/app/constants/api_constants.dart
@@ -1,0 +1,8 @@
+import 'package:flutter_dotenv/flutter_dotenv.dart';
+
+abstract class ApiConstants {
+  static String kBaseUrl = dotenv.env['BASE_URL'] ?? '';
+
+  // user
+  static String kGetUser = 'user_profile';
+}