+++ bricks/project_setup/__brick__/lib/src/app/routing/route_constants.dart
@@ -1,0 +1,4 @@
+class RouteConstants {
+  static const String kSplash = '/splash';
+  static const String kOnBoarding = '/on-boarding';
+}