+++ bricks/project_setup/__brick__/lib/src/app/utilities/typedef.dart
@@ -1,0 +1,7 @@
+import 'package:fpdart/fpdart.dart';
+
+import 'errors/failure.dart';
+
+typedef ResultFuture<T> = Future<Either<Failure, T>>;
+
+typedef DataMap = Map<String, dynamic>;