+++ lib/flutter_blueprint_cli.dart
@@ -8,5 +8,5 @@ export 'src/commands/upgrade_command.dart';
 export 'src/utils/dependency_handler.dart';
 export 'src/utils/project_validator.dart';
 export 'src/utils/generator_helper.dart';
-export 'src/bundles/clean_feature_bundle.dart';
+export 'src/bundles/component_generator_bundle.dart';
 export 'src/bundles/project_setup_bundle.dart';